Moves metadata into core
