--- mnamer/api/__version__.py
@@ -1,1 +1,0 @@
-VERSION = "3.9.0"
