--- mnamer/api/exceptions.py
@@ -1,1 +1,0 @@
-"""Exceptions used by mapi."""
