+++ mnamer/core/metadata.py
